feat(entities): moved side nav to entities, change styles, removed rout new-project
