--- src/app/new-project/page.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { NewProject } from '@/pages/new-project-page';
-
-const Page = () => {
-	return <NewProject />;
-};
-
-export default Page;
