+++ src/entities/side-nav/ui/side-nav.module.scss
