--- src/pages/new-project-page/index.ts
@@ -1,1 +1,0 @@
-export { NewProject } from './ui/new-project-page';
