--- src/pages/new-project-page/ui/new-project-page.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const NewProject = () => {
-	return <div>Новый проект</div>;
-};
