+++ src/shared/layouts/profile-layout/profile-layout.module.scss
@@ -16,13 +16,10 @@
 		line-height: 48px; /* 120% */
 		letter-spacing: -0.4px;
         margin: 0;
-
     }
 }
-
 .container {
 	display: flex;
 	min-height: 968px;
-
 	gap: 40px;
 }
